backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
playwright==1.49.1
groq==0.15.0
supabase==2.11.0
apscheduler==3.10.4
pydantic==2.10.4
httpx==0.28.1
python-dotenv==1.0.1
